+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.2 - 2018-05-16
+### Added
+– CP Clear Cache now remembers your checked boxes
+
+### Fixed
+- Fixes the missing nav item icon on Craft 3.0.7 and above
+
 ## 1.0.1 - 2018-03-28
 ### Fixed
 - Fixes issue #1, where CP Clear Cache would conflict w/ other plugins (e.g. SEOmatic)